Industrial Production Indices in manufacturing in the European Union

A dataset containing on monthly industrial production indices in manufacturing in the European Union (from sts_inpr_m dataset of Eurostat). Data are based 100 in 2015 and are unadjusted, i.e. neither seasonally adjusted nor calendar adjusted.

Usage

ipi_c_eu

ipi_c_eu_df

Format

A monthly ts object from january 1990 to december 2017 with 34 variables for ipi_c_eu and a data.frame for ipi_c_eu_df.

An object of class data.frame with 360 rows and 35 columns.

Source

Eurostat, 'sts_inpr_m' database.

Details

The dataset contains 34 time series corresponding to the following geographical area

BE	Belgium

BG	Bulgaria

CZ	Czechia

DK	Denmark

DE	Germany (until 1990 former territory of the FRG)

EE	Estonia

IE	Ireland

EL	Greece

ES	Spain

FR	France

HR	Croatia

IT	Italy

CY	Cyprus

LV	Latvia

LT	Lithuania

LU	Luxembourg

HU	Hungary

MT	Malta

NL	Netherlands

AT	Austria

PL	Poland

PT	Portugal

RO	Romania

SI	Slovenia

SK	Slovakia

FI	Finland

SE	Sweden

UK	United Kingdom

NO	Norway

CH	Switzerland

ME	Montenegro

MK	Former Yugoslav Republic of Macedonia, the

RS	Serbia

TR	Turkey

BA	Bosnia and Herzegovina